create onCalendarStateChange prop
