+++ package.json
@@ -55,7 +55,6 @@
     "eslint": "^6.3.0",
     "metro-react-native-babel-preset": "^0.56.0",
     "prettier": "^1.18.2",
-    "prop-types": "^15.7.2",
     "typescript": "^5.4.2"
   }
 }
